remove check for state in street selection
